+++ RPG/src/MoveableObject.java
@@ -6,12 +6,12 @@ public class MoveableObject extends GameObject{
 	protected double pastX, pastY; // how much they moved in their last frame
 	protected double right, down; //current direction they're facing
 	private int cooldown = 0;// to move to MO
-	private int loss=0;
+	private double loss=0;
 
 	public MoveableObject(double x, double y, int width, int height , int startingHealth, BufferedImage b) {
 		super(x, y, width, height, false, false, startingHealth, b);
 	}
-	public int getLoss() {
+	public double getLoss() {
 		return loss;
 	}
 	public void update() {
@@ -20,7 +20,7 @@ public class MoveableObject extends GameObject{
 		}
 	}
 
-	public int getDamage() {
+	public double getDamage() {
 		return this.damage;
 	}
 	public double getPastX() {
@@ -68,7 +68,7 @@ public class MoveableObject extends GameObject{
 		pastY=howMuch;
 	}
 	@Override
-	public void hit(int damage) {
+	public void hit(double damage) {
 		if (RPGGame.ticks>super.getHittable()) {
 			//cooldown=26+RPGGame.ticks;
 			loss = damage;
